requirements.txt
fastapi
uvicorn[standard]
python-multipart
perplexityai
jinja2
python-dotenv
